Fixed connection issue + big refactor of the code base
