+++ lib/custom_code/actions/bluetooth_actions.dart
@@ -4,35 +4,23 @@ import 'package:anti_moustique/backend/schema/structs/antimoustique_struct.dart'
 import 'package:anti_moustique/app_state.dart';
 
 class BluetoothActions {
-  static StreamSubscription<List<ScanResult>>? _scanSubscription =
-      FlutterBluePlus.onScanResults.listen((results) {
+  static StreamSubscription<List<ScanResult>>? _scanSubscription = FlutterBluePlus.onScanResults.listen((results) {
     if (results.isNotEmpty) {
       ScanResult r = results.last;
-      print(
-          'Device id : ${r.device.remoteId}, Name : ${r.advertisementData.advName}');
+      print('Device id : ${r.device.remoteId}, Name : ${r.advertisementData.advName}');
     }
   }, onError: (e) => print('Error: $e'));
 
-  StreamSubscription<List<ScanResult>>? get scanSubscription =>
-      _scanSubscription;
-  set scanSubscription(StreamSubscription<List<ScanResult>>? val) =>
-      _scanSubscription = val;
+  StreamSubscription<List<ScanResult>>? get scanSubscription => _scanSubscription;
+  set scanSubscription(StreamSubscription<List<ScanResult>>? val) => _scanSubscription = val;
   bool hasScanSubscription() => _scanSubscription != null;
 
-  static Future<void> scanForDevice(
-      {String manufactureID = '',
-      String remoteID = '',
-      Duration timeout = const Duration(seconds: 1)}) async {
+  static Future<void> scanForDevice({String manufactureID = '', String remoteID = '', Duration timeout = const Duration(seconds: 4)}) async {
     FlutterBluePlus.cancelWhenScanComplete(_scanSubscription!);
 
-    await FlutterBluePlus.adapterState
-        .where((val) => val == BluetoothAdapterState.on)
-        .first;
+    await FlutterBluePlus.adapterState.where((val) => val == BluetoothAdapterState.on).first;
 
-    await FlutterBluePlus.startScan(
-        timeout: timeout,
-        withNames: manufactureID.isNotEmpty ? [manufactureID] : const [],
-        withRemoteIds: remoteID.isNotEmpty ? [remoteID] : const []);
+    await FlutterBluePlus.startScan(timeout: timeout, withNames: manufactureID.isNotEmpty ? [manufactureID] : []);
 
     await FlutterBluePlus.isScanning.where((val) => val == false).first;
   }
@@ -43,36 +31,29 @@ class BluetoothActions {
     print('Connected to device ${antimoustique.device.remoteId}');
   }
 
-  static Future<List<BluetoothService>> discoverServices(
-      AntimoustiqueStruct antimoustique) async {
+  static Future<List<BluetoothService>> discoverServices(AntimoustiqueStruct antimoustique) async {
     var services = await antimoustique.device.discoverServices();
 
     services.forEach((service) {
-      print(
-          'Discovered service ${service.uuid} on device ${antimoustique.device.remoteId}');
+      print('Discovered service ${service.uuid} on device ${antimoustique.device.remoteId}');
     });
 
     return services;
   }
 
-  static Future<List<BluetoothCharacteristic>> discoverCharacteristics(
-      AntimoustiqueStruct antimoustique, BluetoothService service) async {
+  static Future<List<BluetoothCharacteristic>> discoverCharacteristics(AntimoustiqueStruct antimoustique, BluetoothService service) async {
     var characteristics = service.characteristics;
 
     return characteristics;
   }
 
   // Example of a refactored helper method
-  static Future<List<BluetoothCharacteristic>> discoverCharacteristicsForServiceUUID(
-      AntimoustiqueStruct antimoustique, String serviceUUID) async {
+  static Future<List<BluetoothCharacteristic>> discoverCharacteristicsForServiceUUID(AntimoustiqueStruct antimoustique, String serviceUUID) async {
     try {
       return await BluetoothActions.discoverCharacteristics(
-          antimoustique,
-          antimoustique.device.servicesList
-              .firstWhere((service) => service.uuid.toString() == serviceUUID));
+          antimoustique, antimoustique.device.servicesList.firstWhere((service) => service.uuid.toString() == serviceUUID));
     } catch (e) {
-      print(
-          'Error discovering characteristics for service UUID $serviceUUID: $e');
+      print('Error discovering characteristics for service UUID $serviceUUID: $e');
       rethrow; // This allows the calling method to handle the exception
     }
   }
@@ -88,31 +69,24 @@ class BluetoothActions {
         await antimoustique.device.connect(mtu: null);
       }
       await characteristic.write(value, allowLongWrite: allowLongWrite);
-      print(
-          'Wrote to characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
+      print('Wrote to characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
     } catch (e) {
-      throw Exception(
-          'Error writing to characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
+      throw Exception('Error writing to characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
     }
   }
 
-  static Future<List<int>> readFromCharacteristic(
-      AntimoustiqueStruct antimoustique,
-      BluetoothCharacteristic characteristic) async {
+  static Future<List<int>> readFromCharacteristic(AntimoustiqueStruct antimoustique, BluetoothCharacteristic characteristic) async {
     try {
       var value = await characteristic.read();
-      print(
-          'Read from characteristic ${characteristic.uuid}, value : ${value}');
+      print('Read from characteristic ${characteristic.uuid}, value : ${value}');
       return value;
     } catch (e) {
-      print(
-          'Error reading from characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
+      print('Error reading from characteristic ${characteristic.uuid} on device ${antimoustique.device.remoteId}');
       return [];
     }
   }
 
-  static Future<void> disconnectFromDevice(
-      AntimoustiqueStruct antimoustique) async {
+  static Future<void> disconnectFromDevice(AntimoustiqueStruct antimoustique) async {
     await antimoustique.device.disconnect();
 
     print('Disconnected from device ${antimoustique.device.remoteId}');
