--- lib/custom_code/actions/check_if_null.dart
@@ -1,13 +1,0 @@
-// Automatic FlutterFlow imports
-import '/backend/schema/structs/index.dart';
-import '/backend/schema/enums/enums.dart';
-import '/flutter_flow/flutter_flow_theme.dart';
-import '/flutter_flow/flutter_flow_util.dart';
-import 'index.dart'; // Imports other custom actions
-import 'package:flutter/material.dart';
-// Begin custom action code
-// DO NOT REMOVE OR MODIFY THE CODE ABOVE!
-
-Future<bool> checkIfNull(AntimoustiqueStruct? currentDevice) async {
-  return currentDevice == Null;
-}
