+++ lib/custom_code/actions/constants.dart
@@ -1,0 +1,6 @@
+final String deviceInfoServiceUUID = '180d';
+final String deviceCommandServiceUUID = '1900';
+final String functioningScheduleServiceUUID = '6900';
+final String co2LevelCharacteristicUUID = '180e';
+final String attractifLevelCharacteristicUUID = '180f';
+final String activateCommandCharacteristicUUID = '1901';
