--- lib/custom_code/actions/delete_from_list.dart
@@ -1,16 +1,0 @@
-// Automatic FlutterFlow imports
-import '/backend/schema/structs/index.dart';
-import '/backend/schema/enums/enums.dart';
-import '/flutter_flow/flutter_flow_theme.dart';
-import '/flutter_flow/flutter_flow_util.dart';
-import 'index.dart'; // Imports other custom actions
-import 'package:flutter/material.dart';
-// Begin custom action code
-// DO NOT REMOVE OR MODIFY THE CODE ABOVE!
-
-Future deleteFromList(
-  List<AntimoustiqueStruct> deviceList,
-  String? idToDelete,
-) async {
-  deviceList.removeWhere((item) => item.manufactureID == idToDelete);
-}
