--- lib/custom_code/actions/index.dart
@@ -1,2 +1,0 @@
-export 'delete_from_list.dart' show deleteFromList;
-export 'check_if_null.dart' show checkIfNull;
