+++ test/test-device.json
@@ -1,5 +1,5 @@
 {
-    "manufactureID"  : "Pipou" ,
+    "manufactureID"  : "TEST-MOUSTIBOX" ,
     "remoteID" : "07:00:00:00:00:00:00:00" ,
     "vendor" : "MOUSTIBOX"
 }
